Restricted CK set criterion to 2 in S
The refined Chabauty-Kim set in depth 2 is actually empty if 2 is not in S. Consequently, the valuation-theoretic criterion to determine the size of the CK set is restricted to the case S = {2,q}. The function "check_extrapoint_criterion" returns lists of the primes q for which the locus contains an extra point (resp. no extrapoint, resp. cannot be decided), as apposed to lists of pairs of primes (l,q).